src/requirements.txt
pywebview==4.4.1
gevent==23.9.1
gevent-websocket==0.10.1
flask==3.0.0
flask-socketio==5.3.6
flask-apscheduler==1.13.1
mnemonic==0.20
requests==2.31.0
cryptography==41.0.5
ecdsa==0.18.0
pycoin==0.92.20230326
pycryptodome==3.19.0